+++ src/egon/data/datasets/demandregio/__init__.py
@@ -33,14 +33,14 @@ class DemandRegio(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="DemandRegio",
-            version="0.0.1",
+            version="0.0.2",
             dependencies=dependencies,
             tasks=(
                 clone_and_install,
                 create_tables,
+                insert_household_demand,
                 {
                     insert_society_data,
-                    insert_household_demand,
                     insert_cts_ind_demands,
                 },
             ),