+++ src/egon/data/datasets/heat_demand_timeseries/HTS.py
@@ -1,35 +1,35 @@
-import pandas as pd
-import numpy as np
-import psycopg2
-import pandas.io.sql as sqlio
+from datetime import datetime
+import glob
+import os
+
+from sqlalchemy import ARRAY, Column, Float, ForeignKey, Integer, String
+from sqlalchemy.ext.declarative import declarative_base
 import geopandas as gpd
 import matplotlib.pyplot as plt
-import os
-import glob
+import numpy as np
+import pandas as pd
+import pandas.io.sql as sqlio
+import psycopg2
+import xarray as xr
 
 from egon.data import db, subprocess
-
-import xarray as xr
-from sqlalchemy import Column, String, Float, Integer, ForeignKey, ARRAY
 import egon.data.datasets.era5 as era
 
-from sqlalchemy.ext.declarative import declarative_base
-
 try:
-    from disaggregator import config, data, spatial, temporal, plot
+    from disaggregator import config, data, plot, spatial, temporal
 except ImportError as e:
     pass
 
 from math import ceil
 
-
 from egon.data.datasets import Dataset
+import egon
 
 
 class IdpProfiles:
     def __init__(self, df_index, **kwargs):
         index = pd.date_range(
-            pd.datetime(2011, 1, 1, 0), periods=8760, freq="H"
+            datetime(2011, 1, 1, 0), periods=8760, freq="H"
         )
 
         self.df = pd.DataFrame(index=df_index)
@@ -38,7 +38,7 @@ class IdpProfiles:
 
     def get_temperature_interval(self, how="geometric_series"):
         index = pd.date_range(
-            pd.datetime(2011, 1, 1, 0), periods=8760, freq="H"
+            datetime(2011, 1, 1, 0), periods=8760, freq="H"
         )
         """Appoints the corresponding temperature interval to each temperature
         in the temperature vector.
@@ -201,7 +201,7 @@ def temp_interval():
         Hourly temperature intrerval of all 15 TRY Climate station#s temperature profile
 
     """
-    index = pd.date_range(pd.datetime(2011, 1, 1, 0), periods=8760, freq="H")
+    index = pd.date_range(datetime(2011, 1, 1, 0), periods=8760, freq="H")
     temperature_interval = pd.DataFrame()
     temp_profile = temperature_profile_extract()
 
@@ -234,7 +234,7 @@ def idp_pool_generator():
         "household_heat_demand_profiles",
         "household_heat_demand_profiles.hdf5",
     )
-    index = pd.date_range(pd.datetime(2011, 1, 1, 0), periods=8760, freq="H")
+    index = pd.date_range(datetime(2011, 1, 1, 0), periods=8760, freq="H")
 
     sfh = pd.read_hdf(path, key="SFH")
     mfh = pd.read_hdf(path, key="MFH")
@@ -919,7 +919,7 @@ def h_value():
         Extracted from demandlib.
 
     """
-    index = pd.date_range(pd.datetime(2011, 1, 1, 0), periods=8760, freq="H")
+    index = pd.date_range(datetime(2011, 1, 1, 0), periods=8760, freq="H")
 
     a = 3.0469695
 
@@ -978,7 +978,7 @@ def profile_generator(aggregation_level):
             0
     heat_profile_idp : pandas.DataFrame
         if aggreation_level = 'district'
-            heat profiles for every mv grid subst_id
+            heat profiles for every mv grid bus_id
         else
             heat profiles for every zensus_poppulation_id
     """
@@ -1027,13 +1027,14 @@ def profile_generator(aggregation_level):
 
         heat_profile_idp = pd.merge(
             heat_profile_idp,
-            mv_grid_ind["subst_id"],
+            mv_grid_ind["bus_id"],
             left_on=selected_profiles.index,
             right_on=mv_grid_ind.index,
             how="inner",
         )
-        heat_profile_idp.sort_values("subst_id", inplace=True)
-        heat_profile_idp.set_index("subst_id", inplace=True)
+
+        heat_profile_idp.sort_values("bus_id", inplace=True)
+        heat_profile_idp.set_index("bus_id", inplace=True)
         heat_profile_idp.drop("key_0", axis=1, inplace=True)
 
         heat_profile_dist = heat_profile_dist.groupby(
@@ -1082,7 +1083,7 @@ def residential_demand_scale(aggregation_level):
             0
     heat_demand_profile_mv : pandas.DataFrame
         if aggregation ='district'
-            final demand profiles per mv grid subst_id
+            final demand profiles per mv grid bus_id
         else
             0
     heat_demand_profile_zensus : pandas.DataFrame
@@ -1176,26 +1177,25 @@ def residential_demand_scale(aggregation_level):
             mv_grid.index.difference(district_heating.index), :
         ]
         mv_grid_ind = mv_grid_ind.reset_index()
-
         district_grid = pd.merge(
-            mv_grid_ind[["subst_id", "zensus_population_id"]],
+            mv_grid_ind[["bus_id", "zensus_population_id"]],
             annual_demand[["zensus_population_id", "Station", "demand"]],
             on="zensus_population_id",
             how="inner",
         )
 
-        district_grid.sort_values("subst_id", inplace=True)
+        district_grid.sort_values("bus_id", inplace=True)
         district_grid.drop("zensus_population_id", axis=1, inplace=True)
-        district_grid = district_grid.groupby(["subst_id", "Station"]).sum()
+        district_grid = district_grid.groupby(["bus_id", "Station"]).sum()
         district_grid.reset_index("Station", inplace=True)
 
         demand_curves_mv = pd.DataFrame()
         for j in range(len(heat_profile_idp.columns)):
             current_district = heat_profile_idp.iloc[:, j]
-            subst_id = heat_profile_idp.columns[j]
-            station = district_grid[district_grid.index == subst_id][
+            bus_id = heat_profile_idp.columns[j]
+            station = district_grid[district_grid.index == bus_id][
                 "Station"
-            ][subst_id]
+            ][bus_id]
             if type(station) != str:
                 station = station.reset_index()
                 multiple_stations = pd.DataFrame()
@@ -1205,9 +1205,9 @@ def residential_demand_scale(aggregation_level):
                         h[current_station], axis=0
                     )
                 multiple_stations = multiple_stations.sum(axis=1)
-                demand_curves_mv[subst_id] = multiple_stations
+                demand_curves_mv[bus_id] = multiple_stations
             else:
-                demand_curves_mv[subst_id] = current_district.multiply(
+                demand_curves_mv[bus_id] = current_district.multiply(
                     h[station], axis=0
                 )
         demand_curves_mv = demand_curves_mv.apply(lambda x: x / x.sum())
@@ -1225,9 +1225,10 @@ def residential_demand_scale(aggregation_level):
         )
 
         heat_demand_profile_mv.rename(
-            columns={"key_0": "subst_id"}, inplace=True
+
+            columns={"key_0": "bus_id"}, inplace=True
         )
-        heat_demand_profile_mv.set_index("subst_id", inplace=True)
+        heat_demand_profile_mv.set_index("bus_id", inplace=True)
         heat_demand_profile_mv = heat_demand_profile_mv[
             heat_demand_profile_mv.columns[:-1]
         ].multiply(heat_demand_profile_mv.demand, axis=0)
@@ -1418,17 +1419,17 @@ def cts_demand_per_aggregation_level(aggregation_level):
 
         CTS_per_grid = pd.merge(
             CTS_per_zensus,
-            mv_grid_ind[["subst_id", "zensus_population_id"]],
+            mv_grid_ind[["bus_id", "zensus_population_id"]],
             on="zensus_population_id",
             how="inner",
         )
-        CTS_per_grid.set_index("subst_id", inplace=True)
+        CTS_per_grid.set_index("bus_id", inplace=True)
         CTS_per_grid.drop("zensus_population_id", axis=1, inplace=True)
 
         CTS_per_grid = CTS_per_grid.groupby(lambda x: x, axis=0).sum()
         CTS_per_grid = CTS_per_grid.transpose()
         CTS_per_grid = CTS_per_grid.apply(lambda x: x / x.sum())
-        CTS_per_grid.columns.name = "subst_id"
+        CTS_per_grid.columns.name = "bus_id"
         CTS_per_grid.reset_index(drop=True, inplace=True)
 
         CTS_per_zensus = pd.DataFrame()
@@ -1544,13 +1545,13 @@ def CTS_demand_scale(aggregation_level):
 
         CTS_demands_grid = pd.merge(
             demand,
-            mv_grid_ind[["subst_id", "zensus_population_id"]],
+            mv_grid_ind[["bus_id", "zensus_population_id"]],
             on="zensus_population_id",
             how="inner",
         )
 
         CTS_demands_grid.drop("zensus_population_id", axis=1, inplace=True)
-        CTS_demands_grid = CTS_demands_grid.groupby("subst_id").sum()
+        CTS_demands_grid = CTS_demands_grid.groupby("bus_id").sum()
 
         CTS_per_grid = pd.merge(
             CTS_per_grid,
@@ -1560,8 +1561,9 @@ def CTS_demand_scale(aggregation_level):
             left_on=CTS_demands_grid.index,
         )
 
-        CTS_per_grid = CTS_per_grid.rename(columns={"key_0": "subst_id"})
-        CTS_per_grid.set_index("subst_id", inplace=True)
+
+        CTS_per_grid = CTS_per_grid.rename(columns={"key_0": "bus_id"})
+        CTS_per_grid.set_index("bus_id", inplace=True)
 
         CTS_per_grid = CTS_per_grid[CTS_per_grid.columns[:-1]].multiply(
             CTS_per_grid.demand, axis=0
@@ -1643,10 +1645,11 @@ def demand_profile_generator(aggregation_level="district"):
             [residential_demand_grid, CTS_demand_grid]
         )
         total_demands_grid.sort_index(inplace=True)
+
         total_demands_grid = total_demands_grid.groupby(
             lambda x: x, axis=0
         ).sum()
-        total_demands_grid.index.name = "subst_id"
+        total_demands_grid.index.name = "bus_id"
 
         final_heat_profiles_grid = pd.DataFrame(index=total_demands_grid.index)
         final_heat_profiles_grid[
@@ -1660,6 +1663,7 @@ def demand_profile_generator(aggregation_level="district"):
             index=True,
             dtype=ARRAY(Float()),
         )
+
     else:
         total_demands_zensus = pd.concat(
             [residential_demand_zensus, CTS_demand_zensus]
@@ -1691,7 +1695,7 @@ class HeatTimeSeries(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="HeatTimeSeries",
-            version="0.0.0",
+            version="0.0.1",
             dependencies=dependencies,
             tasks=(demand_profile_generator),
         )